🚚 api: certification-attestations-for-session (second route):
move controller to certification course src
